fix kernel.ld

+++ kernel/types.h
@@ -4,7 +4,7 @@ typedef unsigned char  uchar;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int  uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 typedef uint64 pte_t;
 typedef uint64 pde_t;
 typedef uint64 *pagetable_t;// 512 PTEs
